Move templates to dedicated folder, views cleanup
