+++ src/bda/plone/orders/browser/export.py
@@ -12,9 +12,9 @@ from bda.plone.orders import message_factory as _
 from bda.plone.orders import permissions
 from bda.plone.orders import safe_encode
 from bda.plone.orders import safe_filename
-from bda.plone.orders.browser.views import ContentViewBase
-from bda.plone.orders.browser.views import customers_form_vocab
-from bda.plone.orders.browser.views import vendors_form_vocab
+from bda.plone.orders.browser.common import ContentViewBase
+from bda.plone.orders.browser.common import customers_form_vocab
+from bda.plone.orders.browser.common import vendors_form_vocab
 from bda.plone.orders.common import DT_FORMAT
 from bda.plone.orders.common import OrderData
 from bda.plone.orders.common import get_bookings_soup
@@ -137,7 +137,7 @@ def cleanup_for_csv(value):
 
 
 class ExportOrdersForm(YAMLForm, ContentViewBase):
-    browser_template = ViewPageTemplateFile('export.pt')
+    browser_template = ViewPageTemplateFile('templates/export.pt')
     form_template = 'bda.plone.orders.browser:forms/orders_export.yaml'
     message_factory = _
     action_resource = 'exportorders'
