+++ src/bda/plone/orders/browser/mailtemplates.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 from bda.plone.orders import message_factory as _
-from bda.plone.orders.browser.views import ContentViewBase
+from bda.plone.orders.browser.common import ContentViewBase
 from bda.plone.orders.interfaces import IDynamicMailTemplateLibraryStorage
 from bda.plone.orders.mailtemplates import DEFAULT_TEMPLATE_ATTRS
 from bda.plone.orders.mailtemplates import DynamicMailTemplate
